Use uint16_t for port number.

+++ src/irc.cpp
@@ -228,9 +228,9 @@ void ThreadIRCSeed2(void* parg)
 
     while (!fShutdown)
     {
-        CService addrConnect("92.243.23.21", 6667); // irc.lfnet.org
+        CService addrConnect("92.243.23.21", (uint16_t)6667); // irc.lfnet.org
 
-        CService addrIRC("irc.lfnet.org", 6667, true);
+        CService addrIRC("irc.lfnet.org", (uint16_t)6667, true);
         if (addrIRC.IsValid())
             addrConnect = addrIRC;
